+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+
+## 2.1.0
+- plugins: added the InputHelper object and its methods
+- plugins: added new methods
+- plugins: added events
+- plugins: updated config parameters
+- fix xss
+- portals: api v11.5
+
 ## 2.0.0
 - redesign
 - specifying products version